+++ src/domain/interfaces/IBook.java
@@ -1,0 +1,21 @@
+package domain.interfaces;
+
+import domain.model.book.Author;
+
+import java.util.List;
+
+public interface IBook {
+
+    Integer getId();
+
+    String getTitle();
+
+    String getPublisher();
+
+    List<Author> getAuthors();
+
+    Integer getEdition();
+
+    Integer getYearPublication();
+
+}
