+++ src/domain/interfaces/IExemplary.java
@@ -1,0 +1,12 @@
+package domain.interfaces;
+
+public interface IExemplary {
+
+    Integer getBookCode();
+
+    Integer getExemplaryCode();
+
+    Boolean getStatus();
+
+    IBook getBook();
+}
